requirements.txt
# Основные зависимости для агента
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
huggingface_hub==0.20.3
ddgs

# GigaChat
gigachat==0.1.21

# HTTP запросы
requests==2.31.0

# Работа с .env
python-dotenv==1.0.0

# Анализ данных
pandas==2.2.1
numpy==1.26.4
statsmodels==0.14.1

# UI
gradio==4.36.0

# Метрики
rouge-score==0.1.2
scikit-learn==1.5.0

# Для локальной LLM оценки
transformers==4.38.2
torch==2.2.1
accelerate==0.27.2